[refactor] 자바 컨벤션에 맞게 리팩토링
- 문자열, 숫자는 자바 상수로 변환

+++ src/main/java/pay/Pay.java
@@ -1,19 +1,21 @@
 package pay;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import domain.OrderMenu;
 import view.InputView;
 
 public class Pay {
 	private static final double CASH_DISCOUNT = 0.98;
-	private static final double CHICKEN_DISCOUNT = 10000;
+	private static final int CHICKEN_DISCOUNT = 10000;
 	private static final double CREDIT_CARD = 1;
 	private static final double CASH = 2;
 
 	public static void pay(List<OrderMenu> orderMenuList, int tableNumber) {
 		printOrderList(orderMenuList);
 		printTableNumber(tableNumber);
+		// int chickenDiscountMoney = discountMoneyChickenMoreThanTen(orderMenuList);
 		int paymentType = InputView.inputPaymentType();
 		int payMoney = getTotalMoneyToPay(orderMenuList);
 		printPaymentResult(paymentType, payMoney);
@@ -33,6 +35,24 @@ public class Pay {
 		}
 		return payment;
 	}
+
+	// private static int discountMoneyChickenMoreThanTen(List<OrderMenu> orderMenuList) {
+	// 	int totalChickenNumber = getNumberOfChicken(orderMenuList);
+	// 	int discountNumber = totalChickenNumber / 10;
+	// 	System.out.println("치킨 개수" + totalChickenNumber);
+	// 	System.out.println("할인 받을 개수" + discountNumber);
+	// 	if (discountNumber > 0) {
+	// 		System.out.println("할인 받을 금액" + CHICKEN_DISCOUNT * discountNumber);
+	// 		return CHICKEN_DISCOUNT * discountNumber;
+	// 	}
+	// 	return 0;
+	// }
+
+	// private static void getNumberOfChicken(List<OrderMenu> orderMenuList) {
+	// 	return orderMenuList.stream()
+	// 		.filter(orderMenu -> orderMenu.isChicken())
+	// 		.map(orderMenu -> orderMenu.getNumber())
+	// }
 
 	private static int getTotalMoneyToPay(List<OrderMenu> orderMenuList) {
 		return orderMenuList.stream().map(orderMenu -> orderMenu.getTotalPrice()).reduce(Integer::sum).get();
